pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "q2-bias-sim"
version = "0.1.0"
description = "QIIME 2 plugin for applying measurement bias to microbial community proportions"
requires-python = ">=3.8"
dependencies = [
    "qiime2",
    "q2-types",
    "numpy",
    "pandas",
    "biom-format",
    "matplotlib",
    "q2templates",
]

[project.entry-points."qiime2.plugins"]
"q2-bias-sim" = "q2_bias_sim.plugin_setup:plugin"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
q2_bias_sim = ["templates/**/*"]
